requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
influxdb==5.3.1
msgpack==1.0.3
mysql-connector-python==8.0.27
protobuf==3.19.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
